feat: serialize/deserialize logical and execution plan via substrait (#317)
* fix: change Utf8Array indice type


* refactor: remove unused sub-crate


* feat: impl for both Logical and Execution plan


* refactor: move test-util subcrate into table


* test: table scan logical plan round trip


* drop support of physical plan


* fix warnings


* rename trait fns to encode/decode


* address review comments

+++ src/common/grpc/src/physical/plan.rs
@@ -184,7 +184,7 @@ impl ExecutionPlan for MockExecution {
         _runtime: Arc<RuntimeEnv>,
     ) -> datafusion::error::Result<SendableRecordBatchStream> {
         let id_array = Arc::new(PrimitiveArray::from_slice([1u32, 2, 3, 4, 5]));
-        let name_array = Arc::new(Utf8Array::<i64>::from_slice([
+        let name_array = Arc::new(Utf8Array::<i32>::from_slice([
             "zhangsan", "lisi", "wangwu", "Tony", "Mike",
         ]));
         let age_array = Arc::new(PrimitiveArray::from_slice([25u32, 28, 27, 35, 25]));
